refactor: 스프링 설정, 공통 응답 관련 모듈 분리
