+++ src/main/java/com/yeohangttukttak/api/domain/place/entity/PlaceType.java
@@ -1,9 +1,8 @@
 package com.yeohangttukttak.api.domain.place.entity;
 
 import com.fasterxml.jackson.annotation.JsonValue;
-import com.yeohangttukttak.api.domain.ValueBasedEnum;
+import com.yeohangttukttak.api.global.interfaces.ValueBasedEnum;
 import lombok.AllArgsConstructor;
-import lombok.Getter;
 
 @AllArgsConstructor
 public enum PlaceType implements ValueBasedEnum {
